docs: Update build and test instructions in README
The README.md file was updated to reflect the use of the Makefile for building and testing the project. The previous instructions referred to `go build` and `go test`, which have been replaced with the simpler `make build` and `make test` commands.

This change also introduces documentation for the `make release` and `make clean` targets.